Fix frustum culling, PVS, and 64-bit safety issues
- Re-enable R_CullBox (was unconditionally returning false)
- Fix R_SetFrustum: remove incorrect VectorNormalize and -16.0f
  bias on frustum plane dist (restores original GLQuake behavior)
- Change r_novis default from "1" to "0" to enable PVS culling
- Fix 64-bit pointer truncation in NUM_FOR_EDICT, aliashdr offset
  calculations (gl_mesh.c, gl_model.c), and snd_dma.c printf format

+++ src/renderer/gl_mesh.c
@@ -347,12 +347,12 @@ void GL_MakeAliasModelDisplayLists (model_t *m, aliashdr_t *hdr)
 	paliashdr->poseverts = numorder;
 
 	cmds = Hunk_Alloc (numcommands * 4);
-	paliashdr->commands = (byte *)cmds - (byte *)paliashdr;
+	paliashdr->commands = (int)((byte *)cmds - (byte *)paliashdr);
 	memcpy (cmds, commands, numcommands * 4);
 
 	verts = Hunk_Alloc (paliashdr->numposes * paliashdr->poseverts 
 		* sizeof(trivertx_t) );
-	paliashdr->posedata = (byte *)verts - (byte *)paliashdr;
+	paliashdr->posedata = (int)((byte *)verts - (byte *)paliashdr);
 	for (i=0 ; i<paliashdr->numposes ; i++)
 		for (j=0 ; j<numorder ; j++)
 			*verts++ = poseverts[i][vertexorder[j]];
