+++ src/renderer/gl_model.c
@@ -1435,7 +1435,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 			// save 8 bit texels for the player model to remap
 	//		if (!strcmp(loadmodel->name,"progs/player.mdl")) {
 				texels = Hunk_AllocName(s, loadname);
-				pheader->texels[i] = texels - (byte *)pheader;
+				pheader->texels[i] = (int)(texels - (byte *)pheader);
 				memcpy (texels, (byte *)(pskintype + 1), s);
 	//		}
 			sprintf (name, "%s_%i", loadmodel->name, i);
@@ -1460,7 +1460,7 @@ void *Mod_LoadAllSkins (int numskins, daliasskintype_t *pskintype)
 					Mod_FloodFillSkin( skin, pheader->skinwidth, pheader->skinheight );
 					if (j == 0) {
 						texels = Hunk_AllocName(s, loadname);
-						pheader->texels[i] = texels - (byte *)pheader;
+						pheader->texels[i] = (int)(texels - (byte *)pheader);
 						memcpy (texels, (byte *)(pskintype), s);
 					}
 					sprintf (name, "%s_%i_%i", loadmodel->name, i,j);
